requirements.txt
Flask
requests
python-dotenv
PyGithub 
openai